+++ packages/ripple-binary-codec/HISTORY.md
@@ -11,6 +11,12 @@
 * Remove `assert` dependency. If you were catching `AssertionError` you need to change to `Error`.
 * Remove `create-hash` in favor of `@noble/hashes`
 
+### Changes
+* Update type definitions which causing errors in tests that the code already supported
+    * `makeParser` to accept a `Buffer` in addition to `string`
+    * `SerializedType` constructor allows not passing in a byte array
+    * `Comparable` is now a generic type so that it allows `compareTo` methods to take more that the type itself.
+
 ## 1.11.0 (2023-11-30)
 ### Added
 - Support for the DID amendment (XLS-40).